refactoring and tests to customer login
