+++ src/Marcoshoya/MarquejogoBundle/Resources/views/Customer/Secured/login.html.twig
@@ -53,7 +53,7 @@
                 </div>
                 <div class="form-login-group" style="margin: 140px 20px 5px !important">
                     <div class="login-submit">
-                        <input type="submit" class="btn" value="Continuar" />
+                        <input type="submit" class="btn" value="Cadastrar" />
                     </div>
                 </div>
             </fieldset>
